@@ -1,15 +1,15 @@
-import { useState } from "react";
-import Stack from '@mui/material/Stack';
+import AddIcon from "@mui/icons-material/Add";
+import DeleteOutlinedIcon from "@mui/icons-material/DeleteOutlined";
+import HelpOutlineIcon from "@mui/icons-material/HelpOutline";
+import { Box, IconButton } from "@mui/material";
+import FormControlLabel from "@mui/material/FormControlLabel";
+import Stack from "@mui/material/Stack";
+import Tooltip from "@mui/material/Tooltip";
 import Typography from "@mui/material/Typography";
-import Tooltip from '@mui/material/Tooltip';
-import FormControlLabel from '@mui/material/FormControlLabel';
-import HelpOutlineIcon from '@mui/icons-material/HelpOutline';
-import {Box, IconButton} from "@mui/material";
+import { useState } from "react";
 import { vars } from "../../theme/variables.ts";
 import CustomSwitch from "./CustomSwitch.tsx";
 import PickerWrapper from "./PickerWrapper.tsx";
-import DeleteOutlinedIcon from '@mui/icons-material/DeleteOutlined';
-import AddIcon from '@mui/icons-material/Add';
 
 const { gray600, gray400B, gray500, gray50, error700 } = vars;
 
@@ -17,21 +17,21 @@ const CustomListItem = ({ data, showTooltip = true, listType, showExtraActions =
   const [checked, setChecked] = useState(false);
   const [anchorEl, setAnchorEl] = useState(null);
   const [open, setOpen] = useState(false);
-  const [selectedColor, setSelectedColor] = useState('#9FEE9A');
+  const [selectedColor, setSelectedColor] = useState("#9FEE9A");
   const [itemHovered, setItemHovered] = useState(false);
-  
-  const isNeurons = listType === 'neurons';
+
+  const isNeurons = listType === "neurons";
   const onSwitchChange = (e) => {
     setChecked(e.target.checked);
   };
-  
+
   const handleClick = (event) => {
     event.stopPropagation();
     event.preventDefault();
     setAnchorEl(event.currentTarget);
     setOpen(true);
   };
-  
+
   const handleClose = () => {
     setAnchorEl(null);
     setOpen(false);
@@ -39,19 +39,23 @@ const CustomListItem = ({ data, showTooltip = true, listType, showExtraActions =
   const handleColorChange = (color) => {
     setSelectedColor(color.hex);
   };
-  
-    const handleOnMouseEnter = () => {
+
+  const handleOnMouseEnter = () => {
     setItemHovered(true);
   };
-    
-    const handleOnMouseLeave = () => {
+
+  const handleOnMouseLeave = () => {
     setItemHovered(false);
   };
-  
+
   return (
     <>
       <FormControlLabel
-        control={<Tooltip  title={'data.helpText'}><CustomSwitch /></Tooltip>}
+        control={
+          <Tooltip title={"data.helpText"}>
+            <CustomSwitch />
+          </Tooltip>
+        }
         onChange={onSwitchChange}
         onMouseEnter={handleOnMouseEnter}
         onMouseLeave={handleOnMouseLeave}
@@ -59,51 +63,41 @@ const CustomListItem = ({ data, showTooltip = true, listType, showExtraActions =
           m: 1,
           alignItems: "baseline",
           padding: "0.5rem",
-          '&:hover': {
+          "&:hover": {
             background: gray50,
-            borderRadius: '.5rem',
+            borderRadius: ".5rem",
           },
           "& .MuiFormControlLabel-label": {
             width: "100%",
           },
-          
+
           "& .MuiIconButton-root": {
-            padding: '.25rem',
-            borderRadius: '.25rem',
-          }
+            padding: ".25rem",
+            borderRadius: ".25rem",
+          },
         }}
         checked={checked}
         label={
           <Stack>
-            <Stack
-              direction="row"
-              alignItems="center"
-              width={1}
-              spacing=".5rem"
-              justifyContent='space-between'
-            >
-              <Stack
-                direction="row"
-                alignItems="center"
-                justifyContent='space-between'
-                spacing=".5rem"
-              >
-                {isNeurons && <Box
-                  onClick={handleClick}
-                  sx={{
-                    borderRadius: '0.125rem',
-                    border: '1px solid rgba(0, 0, 0, 0.20)',
-                    background: selectedColor,
-                    width: '0.875rem',
-                    height: '0.875rem'
-                  }} />}
+            <Stack direction="row" alignItems="center" width={1} spacing=".5rem" justifyContent="space-between">
+              <Stack direction="row" alignItems="center" justifyContent="space-between" spacing=".5rem">
+                {isNeurons && (
+                  <Box
+                    onClick={handleClick}
+                    sx={{
+                      borderRadius: "0.125rem",
+                      border: "1px solid rgba(0, 0, 0, 0.20)",
+                      background: selectedColor,
+                      width: "0.875rem",
+                      height: "0.875rem",
+                    }}
+                  />
+                )}
                 <Typography color={gray600} variant="subtitle1">
-                  {data?.label?.length > 32
-                    ? data?.label.slice(0, 32) + "..."
-                    : data?.label}
+                  {data?.label?.length > 32 ? `${data?.label.slice(0, 32)}...` : data?.label}
                 </Typography>
               </Stack>
-             
+
               {showTooltip && (
                 <Tooltip title={data.helpText}>
                   <HelpOutlineIcon
@@ -114,28 +108,30 @@ const CustomListItem = ({ data, showTooltip = true, listType, showExtraActions =
                   />
                 </Tooltip>
               )}
-              {
-                showExtraActions && itemHovered && <Box display='flex' alignItems='center' gap='.25rem'>
-                <Tooltip title='Remove from all viewers'>
-                  <IconButton sx={{
-                    padding: '0.125rem !important',
-                    
-                    '&:hover': {
-                      '& .MuiSvgIcon-root': {
-                        color: error700
-                      }
-                    }
-                  }}>
-                    <DeleteOutlinedIcon fontSize='small' />
-                  </IconButton>
-                </Tooltip>
-                  <Tooltip title='Add to group'>
-                  <IconButton>
-                    <AddIcon fontSize='small' />
-                  </IconButton>
+              {showExtraActions && itemHovered && (
+                <Box display="flex" alignItems="center" gap=".25rem">
+                  <Tooltip title="Remove from all viewers">
+                    <IconButton
+                      sx={{
+                        padding: "0.125rem !important",
+
+                        "&:hover": {
+                          "& .MuiSvgIcon-root": {
+                            color: error700,
+                          },
+                        },
+                      }}
+                    >
+                      <DeleteOutlinedIcon fontSize="small" />
+                    </IconButton>
+                  </Tooltip>
+                  <Tooltip title="Add to group">
+                    <IconButton>
+                      <AddIcon fontSize="small" />
+                    </IconButton>
                   </Tooltip>
                 </Box>
-              }
+              )}
             </Stack>
             {data.description && (
               <Typography color={gray500} variant={"caption"}>
@@ -146,13 +142,7 @@ const CustomListItem = ({ data, showTooltip = true, listType, showExtraActions =
         }
         value={undefined}
       />
-      <PickerWrapper
-        onChange={handleColorChange}
-        selectedColor={selectedColor}
-        onClose={handleClose}
-        open={open}
-        anchorEl={anchorEl}
-      />
+      <PickerWrapper onChange={handleColorChange} selectedColor={selectedColor} onClose={handleClose} open={open} anchorEl={anchorEl} />
     </>
   );
 };
