@@ -1,38 +1,36 @@
-import { FormControlLabel, Tooltip, Box, Stack, Typography } from '@mui/material';
+import HelpOutlineIcon from "@mui/icons-material/HelpOutline";
+import { Box, FormControlLabel, Stack, Tooltip, Typography } from "@mui/material";
+import { vars } from "../../../theme/variables.ts"; // Adjust the import path as needed
 import CustomSwitch from "../../ViewerContainer/CustomSwitch.tsx";
-import HelpOutlineIcon from '@mui/icons-material/HelpOutline';
-import {vars} from "../../../theme/variables.ts"; // Adjust the import path as needed
 
-const {gray50, gray600, gray400B} = vars
+const { gray50, gray600, gray400B } = vars;
 const CustomFormControlLabel = ({ label, tooltipTitle, helpText }) => {
   return (
     <FormControlLabel
-      control={<Tooltip title={helpText}><CustomSwitch /></Tooltip>}
+      control={
+        <Tooltip title={helpText}>
+          <CustomSwitch />
+        </Tooltip>
+      }
       sx={{
-        width: '100%',
-        p: '.5rem .5rem .5rem .5rem',
+        width: "100%",
+        p: ".5rem .5rem .5rem .5rem",
         margin: 0,
         alignItems: "baseline",
-        '&:hover': {
+        "&:hover": {
           background: gray50,
-          borderRadius: '.5rem',
+          borderRadius: ".5rem",
         },
         "& .MuiFormControlLabel-label": {
           width: "100%",
         },
         "& .MuiIconButton-root": {
-          borderRadius: '.25rem',
-        }
+          borderRadius: ".25rem",
+        },
       }}
       label={
         <Box>
-          <Stack
-            direction="row"
-            alignItems="center"
-            width={1}
-            spacing=".5rem"
-            justifyContent='space-between'
-          >
+          <Stack direction="row" alignItems="center" width={1} spacing=".5rem" justifyContent="space-between">
             <Typography color={gray600} variant="subtitle1">
               {label}
             </Typography>
