@@ -1,10 +1,10 @@
-import {widgetIds} from "./widgets.ts";
 import ThreeDViewer from "../components/viewers/ThreeD/ThreeDViewer.tsx";
 import TwoDViewer from "../components/viewers/TwoD/TwoDViewer.tsx";
+import { widgetIds } from "./widgets.ts";
 
 const componentMap = {
-    [widgetIds.threeDViewer]: ThreeDViewer,
-    [widgetIds.twoDViewer]: TwoDViewer,
+  [widgetIds.threeDViewer]: ThreeDViewer,
+  [widgetIds.twoDViewer]: TwoDViewer,
 };
 
-export default componentMap
+export default componentMap;
