@@ -1,24 +1,24 @@
 import { WidgetStatus } from "@metacell/geppetto-meta-client/common/layout/model";
 
 export const widgetIds = {
-    threeDViewer: 'threeDViewer',
-    twoDViewer: 'twoDViewer',
-}
+  threeDViewer: "threeDViewer",
+  twoDViewer: "twoDViewer",
+};
 
 export const threeDViewerWidget = () => ({
-    id: widgetIds.threeDViewer,
-    name: "3D Viewer",
-    component: widgetIds.threeDViewer,
-    panelName: "leftPanel",
-    enableClose: false,
-    status: WidgetStatus.ACTIVE,
+  id: widgetIds.threeDViewer,
+  name: "3D Viewer",
+  component: widgetIds.threeDViewer,
+  panelName: "leftPanel",
+  enableClose: false,
+  status: WidgetStatus.ACTIVE,
 });
 
 export const twoDViewerWidget = () => ({
-    id: widgetIds.twoDViewer,
-    name: "Connectivity Graph",
-    component: widgetIds.twoDViewer,
-    panelName: "rightPanel",
-    enableClose: false,
-    status: WidgetStatus.ACTIVE,
+  id: widgetIds.twoDViewer,
+  name: "Connectivity Graph",
+  component: widgetIds.twoDViewer,
+  panelName: "rightPanel",
+  enableClose: false,
+  status: WidgetStatus.ACTIVE,
 });
