@@ -1,35 +1,34 @@
 export enum ViewMode {
-    Default = 'Default',
-    Compare = 'Compare',
+  Default = "Default",
+  Compare = "Compare",
 }
 
 export enum ViewerType {
-    Graph = 'Graph',
-    ThreeD = '3D',
-    EM = 'EM',
-    InstanceDetails = 'Instance Details'
+  Graph = "Graph",
+  ThreeD = "3D",
+  EM = "EM",
+  InstanceDetails = "Instance Details",
 }
 
 export enum ViewerSynchronizationPair {
-    Graph_InstanceDetails,
-    Graph_ThreeD,
-    ThreeD_EM
+  Graph_InstanceDetails = 0,
+  Graph_ThreeD = 1,
+  ThreeD_EM = 2,
 }
 
 export interface NeuronGroup {
-    id: string;
-    name: string;
-    color: string;
-    neurons: Set<string>;
+  id: string;
+  name: string;
+  color: string;
+  neurons: Set<string>;
 }
 
-
 export interface Dataset {
-    name: string;
-    // Add other properties as needed
+  name: string;
+  // Add other properties as needed
 }
 
 export interface Neuron {
-    name: string;
-    // Add other properties as needed
+  name: string;
+  // Add other properties as needed
 }
