@@ -1,2 +1,2 @@
-export const TEMPLATE_ACTIVE_DATASETS = ["white_1986_jsh", "white_1986_n2u"]
-export const TEMPLATE_ACTIVE_NEURONS = ["ASEL", "AIYR"]
+export const TEMPLATE_ACTIVE_DATASETS = ["white_1986_jsh", "white_1986_n2u"];
+export const TEMPLATE_ACTIVE_NEURONS = ["ASEL", "AIYR"];
