@@ -1,5 +1,5 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import react from "@vitejs/plugin-react";
+import { defineConfig } from "vite";
 
 const PORT = 9000;
 
@@ -8,19 +8,19 @@ const defaultConfig = {
   plugins: [react()],
   build: {
     sourcemap: !process.env.NO_SOURCEMAP,
-  }
+  },
   // publicDir: 'src/assets',
-}
+};
 
 export default defineConfig(({ command, mode }) => {
   if (command !== "serve") {
-    return defaultConfig
+    return defaultConfig;
   }
-  const isDev = mode === 'development'
-  const port = process.env.PORT || PORT
-  const theDomain = process.env.DOMAIN || 'http://localhost:8000';
+  const isDev = mode === "development";
+  const port = process.env.PORT || PORT;
+  const theDomain = process.env.DOMAIN || "http://localhost:8000";
 
-  const replaceHost = (uri: string, appName: string) => (uri.includes("visualizer") && uri.replace("visualizer", appName + '.' + theDomain)) || uri
+  const replaceHost = (uri: string, appName: string) => (uri.includes("visualizer") && uri.replace("visualizer", appName + "." + theDomain)) || uri;
 
   return {
     ...defaultConfig,
@@ -31,12 +31,12 @@ export default defineConfig(({ command, mode }) => {
       compress: true,
       port: Number(port),
       proxy: {
-        '/api': {
-          target: replaceHost(theDomain, 'visualizer'),
+        "/api": {
+          target: replaceHost(theDomain, "visualizer"),
           changeOrigin: isDev,
-          secure: !isDev
-        }
-      }
-    }
-  }
-})
+          secure: !isDev,
+        },
+      },
+    },
+  };
+});
